test: 메소드 extension 분리, httpbody 테스트
